Include MT19937 pseudorandom number generator
Include an implementation of MT19937 in file prng.[h-c], the
implementation of this pseudorandom number generator was refered to
the sample implementation in the paper named "Mersenne twister: a
623-dimensionally equidistributed uniform pseudo-random number
generator".
To use it, a none zero seed should be set(the default seed is 1) by
calling function prng_set_seed, then a couple of calls to function
prng_gen_number to get the random numbers.