Include the randomness test for implementation of MT19937
To ensure the randomness of my implementation of MT19937, a randomness
test is introduced in fime randomness_test.c.
By the definition of k-Distribution described in paper "Mersenne
twister: a 623-dimensionally equidistributed uniform pseudo-random
number generator", a pseudorandom sequence is said to be v-bit accuracy
if the count of different patterns of leading v bits of sequence are
the same in a given period. To show that MT19937 can reach 32-bit
accuracy, the following test is designed:
1. Generate a pseudorandom sequence
2. Calculate the 0/1 ratio for each bit
The expected 0/1 ratio is around 1:1, since that the count of 0s and 1s
are the same at a given bit position, if we list all possible
combinations of v bits.
And as the test shows, the 0/1 ration of every bit for all 5 rounds are
around 1:1 as we expected.